master:多个数据库连接

+++ 04_mybatis_xml/src/main/java/com/jason/mybatisxml/mapper/UserMapper.java
@@ -10,7 +10,7 @@ public interface UserMapper {
 
     User getOne(Long id);
 
-    int insert(User user);
+    long insert(User user);
 
     void update(User user);
